Move Pending Restart checking to the StateManager
Previously we move the pending restart checking to the Settings's
onDestroy, but this is not reliable enough especially on device
with big memory where the activity just got paused instead of
destroyed.

Instead of hook it into the settings, we can just hook it into the launcher's
state manager, which will be fired when state is changing (like when user
goes back to home).

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>