PisselBar: Dynamically adjust additional padding for first and last icon
When user has no Google Lens installed, the icon will be gone, and
visually the assistant icon will be shifted to the last icon.
But since the additional padding are still statically defined in the
layout, the bar will looks like poorly aligned.

Yes we can indeed set the container side padding instead, but it'll
affect button boundaries and ruins the ripple effect.
So instead, let's do it programmatically on the layout changes.

Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>